move DbScopeFactory<TCommand> into DbScopeFactory as nested class Cache<TCommand>
