--- tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AdoNet/DbScopeFactory{TCommand}.cs
@@ -1,59 +1,0 @@
-﻿// <copyright file="DbScopeFactory{TCommand}.cs" company="Datadog">
-// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
-// </copyright>
-
-using System;
-using System.Data;
-using Datadog.Trace.Configuration;
-
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AdoNet
-{
-    /// <summary>
-    /// A generic wrapper around <see cref="DbScopeFactory"/> which caches values for each TCommand type.
-    /// </summary>
-    /// <typeparam name="TCommand">The db command type.</typeparam>
-    internal static class DbScopeFactory<TCommand>
-    {
-        private static readonly Type _commandType;
-        private static readonly string _dbTypeName;
-        private static readonly string _operationName;
-        private static readonly IntegrationInfo _integrationInfo;
-
-        static DbScopeFactory()
-        {
-            var commandType = typeof(TCommand);
-
-            if (DbScopeFactory.TryGetIntegrationDetails(commandType.FullName, out var integrationId, out var dbTypeName))
-            {
-                // cache values for this TCommand type
-                _commandType = commandType;
-                _dbTypeName = dbTypeName;
-                _operationName = $"{_dbTypeName}.query";
-                _integrationInfo = IntegrationRegistry.GetIntegrationInfo(integrationId.ToString());
-            }
-        }
-
-        public static Scope CreateDbCommandScope(Tracer tracer, IDbCommand command)
-        {
-            var commandType = command.GetType();
-
-            if (commandType == _commandType)
-            {
-                // use the cached values if command.GetType() == typeof(TCommand)
-                return DbScopeFactory.CreateDbCommandScope(tracer, command, _integrationInfo, _dbTypeName, _operationName);
-            }
-
-            // if command.GetType() != typeof(TCommand), we are probably instrumenting a method
-            // defined in a base class like DbCommand and we can't use the cached values
-            if (DbScopeFactory.TryGetIntegrationDetails(command.GetType().FullName, out var integrationId, out var dbTypeName))
-            {
-                var integration = IntegrationRegistry.GetIntegrationInfo(integrationId.ToString());
-                var operationName = $"{dbTypeName}.query";
-                return DbScopeFactory.CreateDbCommandScope(tracer, command, integration, dbTypeName, operationName);
-            }
-
-            return null;
-        }
-    }
-}
